hybrid-rag-chatbot-backend/requirements.txt
# requirements.txt
# Minimal setup for FastAPI + Groq + Frontend testing

# Core Framework
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
pydantic>=2.5.0,<3.0.0

# Database (SQLite)
aiosqlite>=0.19.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0

# CORS (for frontend connection)
fastapi-cors>=0.0.6,<1.0.0
pandas
openpyxl
google-generativeai
faiss-cpu
fastapi
uvicorn
python-dotenv
